pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "easypanel-mcp"
version = "1.0.0"
description = "Model Context Protocol (MCP) server for EasyPanel - Connect AI agents to your infrastructure"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Danny Maaz", email = "dannymaaz200@gmail.com"}
]
maintainers = [
    {name = "Danny Maaz", email = "dannymaaz200@gmail.com"}
]
keywords = [
    "mcp",
    "easypanel",
    "ai",
    "agent",
    "docker",
    "deployment",
    "infrastructure",
    "devops"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Systems Administration",
    "Topic :: Utilities"
]
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
    "typing-extensions>=4.8.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.5.0"
]

[project.urls]
Homepage = "https://github.com/dannymaaz/easypanel-mcp"
Documentation = "https://dannymaaz.github.io/easypanel-mcp/"
Repository = "https://github.com/dannymaaz/easypanel-mcp"
Issues = "https://github.com/dannymaaz/easypanel-mcp/issues"
Contact = "dannymaaz200@gmail.com"

[project.scripts]
easypanel-mcp = "src.server:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "config*"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = ["E501"]

[tool.ruff.isort]
known-first-party = ["src", "config"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=term-missing"
